Use print_function for Python 3 support
PEP 3105 changed `print` from a statement to a function. In order to use
`print` with Python 3, all references need to be updated to a function.
Python 2, however, raises a `SyntaxError` when treating `print` as a
function. Importing `print_function` from __future__ allows `print` as a
function on all versions of Python greater than or equal to 2.6.